Updated findValue function to take pointer to output location. Implemeneted override to output operator to be able to output a binary tree

+++ BinaryTree/BTree.cpp
@@ -1,5 +1,5 @@
 #include "BTree.h"
-#include <iostream>
+
 
 BTree::BTree() {
 	root = nullptr;
@@ -8,6 +8,8 @@ BTree::~BTree() {
 	destroyTree();
 }
 void BTree::insertValue(int value, node* subNode) {
+	if (subNode->value == value)
+		return;
 	bool left = value < subNode->value;
 
 	if (!subNode->leftNode && !subNode->rightNode) {
@@ -47,22 +49,29 @@ void BTree::insertValue(int value) {
 }
 
 node* BTree::findValue(int value, node* subNode) {
-	bool left = value < subNode->value;
 	if (value == subNode->value)
 		return subNode;
+	bool left = value < subNode->value;
 	if (left && subNode->leftNode)
 		return findValue(value, subNode->leftNode);
 	
 	if (!left && subNode->rightNode)
 		return findValue(value, subNode->rightNode);
 	else {
-		node* leaf = new node;
-		return leaf;
+		
+		return nullptr;
 	}
 }
 
-bool BTree::findValue(int value) {
-	return findValue(value, root)->value;
+bool BTree::findValue(int value, node** outNode) {
+	if (!outNode)
+		return false;
+	if (!root)
+		return false;
+	*outNode = findValue(value, root);
+	if (!*outNode)
+		return false;
+	return true;
 }
 void BTree::destroyTree(node* subNode) {
 	if (subNode) {
@@ -72,9 +81,15 @@ void BTree::destroyTree(node* subNode) {
 	}
 }
 void BTree::destroyTree() {
+	if (!root)
+		return;
 	destroyTree(root);
 }
+void BTree::deleteValue(int value) {
+	deleteValue(value, findValue(value, root));
+}
 void BTree::deleteValue(int value, node* subNode) {
+	
 	if (value == subNode->value) {
 		node* minNode = minNodeFromGivenNode(subNode->rightNode);
 		if (minNode->rightNode)
@@ -82,11 +97,41 @@ void BTree::deleteValue(int value, node* subNode) {
 		subNode->value = minNode->value;
 		delete minNode;
 	}
-	//find value search
+	return;
+	
 
 }
 node* BTree::minNodeFromGivenNode(node* subNode) {
 	if (!subNode->leftNode)
 		return subNode;
 	minNodeFromGivenNode(subNode->leftNode);
+}
+void BTree::print() {
+	if (root) {
+		print(root);
+		std::cout << "\n" << std::endl;
+		return;
+	}
+	std::cout << "Error! No tree exists" << std::endl;
+}
+void BTree::print(node* subNode) {
+	if (subNode) {
+		print(subNode->leftNode);
+		std::cout << subNode->value << ", ";
+		print(subNode->rightNode);
+	}
+	return;
+}
+std::ostream& operator<< (std::ostream& os, const BTree& tree) {
+	os << tree.root;
+	return os;
+}
+std::ostream& operator<< (std::ostream& os, const node* node) {
+	if (node->leftNode)
+		os << node->leftNode;
+	os << node->value << ", ";
+	if (node->rightNode)
+		os << node->rightNode;
+
+	return os;
 }
